refactor(utils): move url code to core.url_utils
