+++ ggshield/verticals/auth/oauth.py
@@ -21,7 +21,7 @@ from ggshield.core.client import (
 from ggshield.core.config import Config, InstanceConfig
 from ggshield.core.errors import APIKeyCheckError, UnexpectedError
 from ggshield.core.text_utils import get_pretty_date
-from ggshield.core.utils import urljoin
+from ggshield.core.url_utils import urljoin
 
 
 CLIENT_ID = "ggshield_oauth"
