+++ tests/unit/core/test_utils.py
@@ -5,7 +5,6 @@ from unittest.mock import Mock, patch
 
 import click
 import pytest
-from click import UsageError
 from pygitguardian import GGClient
 
 from ggshield.core.cache import Cache
@@ -16,8 +15,6 @@ from ggshield.core.scan import Commit, Files, ScanContext, ScanMode, StringScann
 from ggshield.core.scan.scan_context import parse_os_release
 from ggshield.core.utils import (
     MatchIndices,
-    api_to_dashboard_url,
-    dashboard_to_api_url,
     find_match_indices,
     get_lines_from_content,
     load_dot_env,
@@ -174,80 +171,6 @@ def test_retrieve_client_unknown_custom_dashboard_url(isolated_fs):
             config = Config()
             config.cmdline_instance_name = "https://example.com"
             create_client_from_config(config)
-
-
-class TestAPIDashboardURL:
-    @pytest.mark.parametrize(
-        ["api_url", "dashboard_url"],
-        [
-            ["https://api.gitguardian.com", "https://dashboard.gitguardian.com"],
-            ["https://api.gitguardian.com/", "https://dashboard.gitguardian.com"],
-            ["https://api.gitguardian.com/v1", "https://dashboard.gitguardian.com"],
-            [
-                "https://api.gitguardian.com/?foo=bar",
-                "https://dashboard.gitguardian.com?foo=bar",
-            ],
-            ["https://example.com/exposed", "https://example.com"],
-            ["https://example.com/exposed/", "https://example.com"],
-            [
-                "https://example.com/exposed/?foo=bar",
-                "https://example.com?foo=bar",
-            ],
-            [
-                "https://example.com/toto/exposed/?foo=bar",
-                "https://example.com/toto?foo=bar",
-            ],
-            [
-                "https://example.com/exposed/v1/?foo=bar",
-                "https://example.com?foo=bar",
-            ],
-        ],
-    )
-    def test_api_to_dashboard_url(self, api_url, dashboard_url):
-        assert api_to_dashboard_url(api_url) == dashboard_url
-
-    @pytest.mark.parametrize(
-        ["dashboard_url", "api_url"],
-        [
-            ["https://dashboard.gitguardian.com", "https://api.gitguardian.com"],
-            ["https://dashboard.gitguardian.com/", "https://api.gitguardian.com"],
-            [
-                "https://dashboard.gitguardian.com/?foo=bar",
-                "https://api.gitguardian.com?foo=bar",
-            ],
-            ["https://example.com/", "https://example.com/exposed"],
-            ["https://example.com/", "https://example.com/exposed"],
-            [
-                "https://example.com/?foo=bar",
-                "https://example.com/exposed?foo=bar",
-            ],
-            [
-                "https://example.com/toto?foo=bar",
-                "https://example.com/toto/exposed?foo=bar",
-            ],
-        ],
-    )
-    def test_dashboard_to_api_url(self, dashboard_url, api_url):
-        assert dashboard_to_api_url(dashboard_url) == api_url
-
-    @pytest.mark.parametrize(
-        "api_url",
-        ["https://api.gitguardian.com/exposed", "https://api.gitguardian.com/toto"],
-    )
-    def test_unexpected_path_api_url(self, api_url):
-        with pytest.raises(UsageError, match="got an unexpected path"):
-            api_to_dashboard_url(api_url)
-
-    @pytest.mark.parametrize(
-        "dashboard_url",
-        [
-            "https://dashboard.gitguardian.com/exposed",
-            "https://dashboard.gitguardian.com/toto",
-        ],
-    )
-    def test_unexpected_path_dashboard_url(self, dashboard_url):
-        with pytest.raises(UsageError, match="got an unexpected path"):
-            api_to_dashboard_url(dashboard_url)
 
 
 @patch("ggshield.core.utils.load_dotenv")
